start working on rust build script for ffmpeg

+++ static-ffmpeg.sh
@@ -2,19 +2,27 @@
 
 set -o errexit
 
-brew install git doxygen asciidoc wget cmake autoconf automake nasm libtool ninja meson pkg-config rtmpdump libtool
+# brew install git doxygen asciidoc wget cmake autoconf automake nasm libtool ninja meson pkg-config rtmpdump
+
+sudo apt-get update
+sudo apt-get upgrade -y git docbook2x asciidoc autopoint wget cmake autoconf automake nasm libtool ninja-build meson pkg-config rtmpdump gperf ragel gtk-doc-tools
 
 # THREADS=$(sysctl -n hw.ncpu)
-THREADS=$(sysctl -n hw.ncpu)
+THREADS=16
 TARGET=$(realpath "./ffmpeg-build")
 CMPL="$TARGET/build"
 mkdir -p "$CMPL"
 # rm -fr $CMPL/*
 export PATH="${TARGET}/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/include:/usr/local/opt:/usr/local/Cellar:/usr/local/lib:/usr/local/share:/usr/local/etc"
 
-if true; then
+export LDFLAGS="-L${TARGET}/lib,-Wl"
+export CPPFLAGS="-I${TARGET}/include,-Wl"
+export CFLAGS="-I${TARGET}/include,-Wl,-fno-stack-check"
+
+if false; then
   # xz
   cd ${CMPL}
+  rm -rf xz
   git -c http.sslVerify=false clone https://git.tukaani.org/xz.git
   cd xz
   ./autogen.sh || (echo "" > /dev/null)
@@ -22,9 +30,9 @@ if true; then
   make -j "$THREADS" && make install
   rm -fr $CMPL/*
 
-
   # libexpat
   cd ${CMPL}
+  rm -rf libexpat/expat
   git clone https://github.com/libexpat/libexpat.git libexpat
   cd libexpat/expat
   ./buildconf.sh
@@ -34,26 +42,35 @@ if true; then
 
   # iconv
   cd ${CMPL}
+  rm -f libiconv*
+  rm -rf libiconv*
   wget --no-check-certificate "https://ftp.gnu.org/pub/gnu/libiconv/libiconv-1.16.tar.gz"
   tar -zxvf libiconv*
   cd libiconv*/
   ./configure --prefix=${TARGET} --enable-static --enable-extra-encodings
   make -j "$THREADS" && make install
   rm -fr $CMPL/*
+fi
 
+if false; then
   # gettext - Requirement for fontconfig, fribidi
   cd ${CMPL}
+  rm -f gettex*
+  rm -rf gettex*
   wget --no-check-certificate "https://ftp.gnu.org/pub/gnu/gettext/gettext-0.21.tar.gz"
   tar -zxvf gettex*
   cd gettex*/
   ./configure --prefix=${TARGET} --disable-dependency-tracking --disable-silent-rules --disable-debug --with-included-gettext --with-included-glib \
-   --with-included-libcroco --with-included-libunistring --with-included-libxml --with-emacs --disable-java --disable-native-java --disable-csharp \
-   --disable-shared --enable-static --without-git --without-cvs --disable-docs --disable-examples
+    --with-included-libcroco --with-included-libunistring --with-included-libxml --with-emacs --disable-java --disable-native-java --disable-csharp \
+    --disable-shared --enable-static --without-git --without-cvs --disable-docs --disable-examples
   make -j "$THREADS" && make install
   rm -fr $CMPL/*
+fi
 
+if false; then
   # libpng git - Requirement for freetype
   cd ${CMPL}
+  rm -rf libpng
   git clone https://github.com/glennrp/libpng.git
   cd libpng
   autoreconf -fiv
@@ -64,24 +81,31 @@ if true; then
   # pkg-config
   cd ${CMPL}
   LastVersion=$(wget --no-check-certificate 'https://pkg-config.freedesktop.org/releases/' -O- -q | grep -Eo 'pkg-config-0.29[0-9\.]+\.tar.gz' | tail -1)
+  rm -f pkg-config-*
+  rm -rf pkg-config-*
   wget --no-check-certificate 'https://pkg-config.freedesktop.org/releases/'"$LastVersion"
   tar -zxvf pkg-config-*
   cd pkg-config-*/
   ./configure --prefix=${TARGET} --disable-debug --disable-host-tool --with-internal-glib
   make -j "$THREADS" && make check && make install
   rm -fr $CMPL/*
+fi
 
+if false; then
   # Yasm
   LastVersion=$(wget --no-check-certificate 'http://www.tortall.net/projects/yasm/releases/' -O- -q | grep -Eo 'yasm-[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f yasm-*
+  rm -rf yasm-*/
   wget --no-check-certificate 'http://www.tortall.net/projects/yasm/releases/'"$LastVersion"
-  tar -zxvf $CMPL/yasm-*
+  tar -zxvf yasm-*
   cd yasm-*/
   ./configure --prefix=${TARGET} && make -j "$THREADS" && make install
   rm -fr $CMPL/*
 
   # bzip2
   cd ${CMPL}
+  rm -rf bzip2
   git clone git://sourceware.org/git/bzip2.git bzip2
   cd bzip2
   make -j "$THREADS" && make install PREFIX=${TARGET}
@@ -91,6 +115,8 @@ if true; then
   # requires correct iconv
   # CPPFLAGS='-I/opt/local/include' LDFLAGS='-L/opt/local/lib'
   cd ${CMPL}
+  rm -f SDL2-*.tar.gz
+  rm -rf SDL2*
   wget http://www.libsdl.org/release/SDL2-2.0.14.tar.gz
   tar xvf SDL2-*.tar.gz
   cd SDL2*/
@@ -101,6 +127,7 @@ if true; then
 
   # libudfread git
   cd ${CMPL}
+  rm -rf libud*
   git clone https://github.com/vlc-mirror/libudfread.git
   cd libud*/
   ./bootstrap
@@ -111,6 +138,7 @@ if true; then
   JAVAV=$(find /Library/Java/JavaVirtualMachines -iname "*.jdk" | tail -1)
   export JAVA_HOME="$JAVAV/Contents/Home"
   cd ${CMPL}
+  rm -rf libblura*
   git -c http.sslVerify=false clone https://code.videolan.org/videolan/libbluray.git
   cd libblura*/
   cp -r $CMPL/libudfread/src $CMPL/libbluray/contrib/libudfread/src
@@ -119,10 +147,27 @@ if true; then
   cp -vpfr $CMPL/libblura*/jni/darwin/jni_md.h $CMPL/libblura*/jni
   make -j "$THREADS" && make install
   rm -fr $CMPL/*
+fi
+
+if true; then
+  # harfbuzz
+  cd ${CMPL}
+  rm -rf harfbuzz
+  git clone https://github.com/harfbuzz/harfbuzz.git
+  cd harfbuzz
+  # meson --prefix=${TARGET} build --buildtype release --default-library static
+  # meson compile -C build
+  ./autogen.sh
+  ./configure --prefix=${TARGET} --disable-shared --enable-static
+  make -j "$THREADS" && make install
+  rm -fr $CMPL/*
+  exit 0
 
   # freetype
   LastVersion=$(wget --no-check-certificate 'https://download.savannah.gnu.org/releases/freetype/' -O- -q | grep -Eo 'freetype-[0-9\.]+\.10+\.[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f freetype-*
+  rm -rf freetype-*
   wget --no-check-certificate 'https://download.savannah.gnu.org/releases/freetype/'"$LastVersion"
   tar xzpf freetype-*
   cd freetype-*/
@@ -133,6 +178,8 @@ if true; then
 
   # fribidi
   cd ${CMPL}
+  rm -f fribid*
+  rm -rf fribid*
   wget --no-check-certificate -O fribidi-1.0.10.tar.xz https://github.com/fribidi/fribidi/releases/download/v1.0.10/fribidi-1.0.10.tar.xz
   tar -xJf fribid*
   cd fribid*/
@@ -142,25 +189,20 @@ if true; then
 
   # fontconfig
   cd ${CMPL}
+  rm -f fontconfig-*
+  rm -rf fontconfig-*
   wget --no-check-certificate https://www.freedesktop.org/software/fontconfig/release/fontconfig-2.13.92.tar.gz
   tar xzpf fontconfig-*
   cd fontconfig-*/
-  ./configure --prefix=${TARGET} --disable-dependency-tracking --disable-silent-rules --with-add-fonts="/System/Library/Fonts,/Library/Fonts" --disable-shared --enable-static
+  ./configure --prefix=${TARGET} --disable-dependency-tracking --disable-silent-rules --with-add-fonts="/System/Library/Fonts,/Library/Fonts" --disable-docs --disable-shared --enable-static
   make -j "$THREADS" && make install
   rm -fr $CMPL/*
-
-  # harfbuzz
-  cd ${CMPL}
-  git clone https://github.com/harfbuzz/harfbuzz.git
-  cd harfbuzz
-  ./autogen.sh
-  ./configure --prefix=${TARGET} --disable-shared --enable-static
-  make -j "$THREADS" && make install
-  rm -fr $CMPL/*
+  exit 0
 
   # libass
   # requires harfbuzz
   cd ${CMPL}
+  rm -rf libas*
   git clone https://github.com/libass/libass.git
   cd libas*/
   ./autogen.sh
@@ -171,6 +213,8 @@ if true; then
   # openssl
   LastVersion=$(wget --no-check-certificate 'https://www.openssl.org/source/' -O- -q | grep -Eo 'openssl-[0-9\.]+\.[0-9\.]+\.[0-9\.]+[A-Za-z].tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f openssl*
+  rm -rf openssl-*
   wget --no-check-certificate https://www.openssl.org/source/"$LastVersion"
   tar -zxvf openssl*
   cd openssl-*/
@@ -181,6 +225,7 @@ if true; then
   # srt
   # requires openssl
   cd ${CMPL}
+  rm -rf srt/
   git clone --depth 1 https://github.com/Haivision/srt.git
   cd srt/
   mkdir build && cd build
@@ -190,6 +235,8 @@ if true; then
 
   # snappy
   cd ${CMPL}
+  rm -f snappy.tar.gz
+  rm -rf snappy-*
   wget -O snappy.tar.gz --no-check-certificate https://github.com/google/snappy/archive/1.1.8.tar.gz
   tar -zxvf snappy.tar.gz
   cd snappy-*/
@@ -200,6 +247,7 @@ if true; then
 
   # openal-soft
   cd ${CMPL}
+  rm -rf openal-soft*
   git clone https://github.com/kcat/openal-soft
   cd openal-soft*/
   cmake -G "Ninja" -DCMAKE_INSTALL_PREFIX:PATH=${TARGET} -DLIBTYPE=STATIC .
@@ -208,6 +256,8 @@ if true; then
 
   # opencore-amr
   cd ${CMPL}
+  rm -f opencore-amr-*
+  rm -rf opencore-amr-*
   wget -O opencore-amr-0.1.5.tar.gz http://freefr.dl.sourceforge.net/project/opencore-amr/opencore-amr/opencore-amr-0.1.5.tar.gz
   tar -zxvf opencore-amr-*
   cd opencore-amr-*
@@ -218,6 +268,8 @@ if true; then
   # opus - Replace speex
   LastVersion=$(wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/opus/ -O- -q | grep -Eo 'opus-1.[0-9\.]+\.[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f opus-*
+  rm -rf opus-*
   wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/opus/"$LastVersion"
   tar -zxvf opus-*
   cd opus-*/
@@ -228,6 +280,8 @@ if true; then
   # ogg
   LastVersion=$(wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/ogg/ -O- -q | grep -Eo 'libogg-[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f libogg-*
+  rm -rf libogg-*
   wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/ogg/"$LastVersion"
   tar -zxvf libogg-*
   cd libogg-*/
@@ -238,6 +292,7 @@ if true; then
   # Theora
   # requires nf automake libtool
   cd ${CMPL}
+  rm -rf theora
   git clone https://github.com/xiph/theora.git
   cd theora
   ./autogen.sh
@@ -248,6 +303,8 @@ if true; then
   # vorbis
   LastVersion=$(wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/vorbis/ -O- -q | grep -Eo 'libvorbis-[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f libvorbis-*
+  rm -rf libvorbis-*
   wget --no-check-certificate https://ftp.osuosl.org/pub/xiph/releases/vorbis/"$LastVersion"
   tar -zxvf libvorbis-*
   cd libvorbis-*/
@@ -257,6 +314,7 @@ if true; then
 
   # lame
   cd ${CMPL}
+  rm -rf lam*
   git clone https://github.com/rbrito/lame.git
   cd lam*/
   ./configure --prefix=${TARGET} --disable-shared --enable-static
@@ -268,6 +326,7 @@ if true; then
   # LastVersion=$(wget --no-check-certificate 'http://www.twolame.org' -O- -q | grep -Eo 'twolame-[0-9\.]+\.tar.gz' | tail -1)
   # wget --no-check-certificate -O twolame.tar.gz 'http://downloads.sourceforge.net/twolame/'"$LastVersion"
   # tar -zxvf twolame.tar.gz
+  rm -rf twolam*
   git clone https://github.com/njh/twolame.git
   cd twolam*/
   ./autogen.sh
@@ -277,6 +336,8 @@ if true; then
 
   # fdk-aac
   cd ${CMPL}
+  rm -f fdk-aac-*
+  rm -rf fdk*
   wget -O fdk-aac-2.0.1.tar.gz http://freefr.dl.sourceforge.net/project/opencore-amr/fdk-aac/fdk-aac-2.0.1.tar.gz
   tar -zxvf fdk-aac-*
   cd fdk*/
@@ -286,6 +347,8 @@ if true; then
 
   # gsm
   cd ${CMPL}
+  rm -f gsm*
+  rm -rf gsm*
   wget --no-check-certificate 'http://www.quut.com/gsm/gsm-1.0.19.tar.gz'
   tar -zxvf gsm*
   cd gsm*/
@@ -301,6 +364,8 @@ if true; then
 
   # speex
   cd ${CMPL}
+  rm -f speex-1.2.0.tar.gz
+  rm -rf speex-1.2.0
   wget http://downloads.us.xiph.org/releases/speex/speex-1.2.0.tar.gz
   tar xvf speex-1.2.0.tar.gz
   cd speex-1.2.0
@@ -310,6 +375,7 @@ if true; then
 
   # libzimg
   cd ${CMPL}
+  rm -rf zimg
   git clone https://github.com/sekrit-twc/zimg.git
   cd zimg
   ./autogen.sh
@@ -319,6 +385,7 @@ if true; then
 
   # libvpx
   cd ${CMPL}
+  rm -rf libvp*
   git clone https://github.com/webmproject/libvpx.git
   cd libvp*/
   ./configure --prefix=${TARGET} --enable-vp8 --enable-postproc --enable-vp9-postproc --enable-vp9-highbitdepth --disable-examples --disable-docs --enable-multi-res-encoding --disable-unit-tests --enable-pic --disable-shared
@@ -327,6 +394,7 @@ if true; then
 
   # webp
   cd ${CMPL}
+  rm -rf libweb*
   git clone https://chromium.googlesource.com/webm/libwebp
   cd libweb*/
   ./autogen.sh
@@ -336,6 +404,7 @@ if true; then
 
   # openjpeg
   cd ${CMPL}
+  rm -rf openjpeg
   git clone https://github.com/uclouvain/openjpeg.git
   cd openjpeg
   mkdir build && cd build
@@ -345,6 +414,7 @@ if true; then
 
   # av1
   cd ${CMPL}
+  rm -rf aom
   git clone https://aomedia.googlesource.com/aom
   cd aom
   mkdir aom_build && cd aom_build
@@ -354,6 +424,7 @@ if true; then
 
   # dav1d git - Require ninja, meson
   cd ${CMPL}
+  rm -rf dav1*
   git -c http.sslVerify=false clone https://code.videolan.org/videolan/dav1d.git
   cd dav1*/
   meson --prefix=${TARGET} build --buildtype release --default-library static
@@ -363,6 +434,8 @@ if true; then
   # xvid
   LastVersion=$(wget --no-check-certificate https://downloads.xvid.com/downloads/ -O- -q | grep -Eo 'xvidcore-[0-9\.]+\.tar.gz' | tail -1)
   cd ${CMPL}
+  rm -f xvidcore*
+  rm -rf xvidcore*
   wget --no-check-certificate https://downloads.xvid.com/downloads/"$LastVersion"
   tar -zxvf xvidcore*
   cd xvidcore/build/generic/
@@ -373,6 +446,7 @@ if true; then
 
   # openh264
   cd ${CMPL}
+  rm -rf openh264/
   git clone https://github.com/cisco/openh264.git
   cd openh264/
   make -j "$THREADS" install-static PREFIX=${TARGET}
@@ -380,6 +454,7 @@ if true; then
 
   # x264 8-10bit git - Require nasm
   cd ${CMPL}
+  rm -rf x264/
   git -c http.sslVerify=false clone https://code.videolan.org/videolan/x264.git
   cd x264/
   ./configure --prefix=${TARGET} --enable-static --bit-depth=all --chroma-format=all --enable-mp4-output
@@ -388,6 +463,7 @@ if true; then
 
   # x265 8-10-12bit - Require wget, cmake, yasm, nasm, libtool, ninja
   cd ${CMPL}
+  rm -rf x265*
   git clone https://bitbucket.org/multicoreware/x265_git/src/master/ x265-master
   cd x265*/source/
   mkdir -p 8bit 10bit 12bit
@@ -415,6 +491,7 @@ if true; then
 
   #_ AviSynth+
   cd ${CMPL}
+  rm -rf AviSynthPlus
   git clone https://github.com/AviSynth/AviSynthPlus.git
   cd AviSynthPlus
   mkdir avisynth-build && cd avisynth-build
@@ -423,19 +500,13 @@ if true; then
   rm -fr $CMPL/*
 fi
 
-# librtmp
-# cp -v /usr/local/Cellar/rtmpdump/2.4+20151223_1/bin/* /Volumes/RamDisk/sw/bin/
-# cp -vr /usr/local/Cellar/rtmpdump/2.4+20151223_1/include/* /Volumes/RamDisk/sw/include/
-# cp -v /usr/local/Cellar/rtmpdump/2.4+20151223_1/lib/pkgconfig/librtmp.pc /Volumes/RamDisk/sw/lib/pkgconfig
-# cp -v /usr/local/Cellar/rtmpdump/2.4+20151223_1/lib/librtmp* /Volumes/RamDisk/sw/lib
-
-export LDFLAGS="-L${TARGET}/lib -Wl,-framework,OpenAL"
-export CPPFLAGS="-I${TARGET}/include -Wl,-framework,OpenAL"
-export CFLAGS="-I${TARGET}/include -Wl,-framework,OpenAL,-fno-stack-check"
+# export LDFLAGS="-L${TARGET}/lib -Wl,-framework,OpenAL"
+# export CPPFLAGS="-I${TARGET}/include -Wl,-framework,OpenAL"
+# export CFLAGS="-I${TARGET}/include -Wl,-framework,OpenAL,-fno-stack-check"
 
 # Minimal ffmpeg
 cd ${CMPL}
-rm -fr ${CMPL}/ffmpeg
+rm -fr ffmpe*
 git clone git://git.ffmpeg.org/ffmpeg.git
 cd ffmpe*/
 ./configure \
@@ -468,15 +539,15 @@ rm -fr ${CMPL}/ffmpeg
 git clone git://git.ffmpeg.org/ffmpeg.git
 cd ffmpe*/
 ./configure --extra-version=adam-"$(date +"%Y-%m-%d")" --extra-cflags="-fno-stack-check" --arch=x86_64 --cc=/usr/bin/clang \
- --enable-pthreads --enable-postproc --enable-runtime-cpudetect \
- --pkg_config='pkg-config --static' --enable-nonfree --enable-gpl --enable-version3 --prefix=${TARGET} \
- --disable-ffplay --disable-ffprobe --disable-debug --disable-doc --enable-avfilter --enable-avisynth --enable-filters \
- --enable-libopus --enable-libvorbis --enable-libtheora --enable-libspeex --enable-libmp3lame --enable-libfdk-aac --enable-encoder=aac \
- --enable-libopencore_amrwb --enable-libopencore_amrnb --enable-libopencore_amrwb --enable-libgsm \
- --enable-muxer=mp4 --enable-libxvid --enable-libopenh264 --enable-libx264 --enable-libx265 --enable-libvpx --enable-libaom --enable-libdav1d \
- --enable-fontconfig --enable-libfreetype --enable-libfribidi --enable-libass --enable-libsrt \
- --enable-libbluray --enable-bzlib --enable-zlib --enable-lzma --enable-libsnappy --enable-libwebp --enable-libopenjpeg \
- --enable-opengl --enable-opencl --enable-openal --enable-libzimg --enable-openssl 
+  --enable-pthreads --enable-postproc --enable-runtime-cpudetect \
+  --pkg_config='pkg-config --static' --enable-nonfree --enable-gpl --enable-version3 --prefix=${TARGET} \
+  --disable-ffplay --disable-ffprobe --disable-debug --disable-doc --enable-avfilter --enable-avisynth --enable-filters \
+  --enable-libopus --enable-libvorbis --enable-libtheora --enable-libspeex --enable-libmp3lame --enable-libfdk-aac --enable-encoder=aac \
+  --enable-libopencore_amrwb --enable-libopencore_amrnb --enable-libopencore_amrwb --enable-libgsm \
+  --enable-muxer=mp4 --enable-libxvid --enable-libopenh264 --enable-libx264 --enable-libx265 --enable-libvpx --enable-libaom --enable-libdav1d \
+  --enable-fontconfig --enable-libfreetype --enable-libfribidi --enable-libass --enable-libsrt \
+  --enable-libbluray --enable-bzlib --enable-zlib --enable-lzma --enable-libsnappy --enable-libwebp --enable-libopenjpeg \
+  --enable-opengl --enable-opencl --enable-openal --enable-libzimg --enable-openssl
 
 # --enable-libtwolame
 # --enable-librtmp
@@ -484,7 +555,8 @@ make -j "$THREADS" && make install
 
 # Check Static
 otool -L $TARGET/bin/ffmpeg
-if otool -L $TARGET/bin/ffmpeg | grep /usr/local
-then echo FFmpeg build Not Static, Please Report
-else echo FFmpeg build Static, Have Fun
+if otool -L $TARGET/bin/ffmpeg | grep /usr/local; then
+  echo FFmpeg build Not Static, Please Report
+else
+  echo FFmpeg build Static, Have Fun
 fi
